requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0
openai==1.6.1
anthropic==0.8.1
python-dotenv==1.0.0
faker==21.0.0
numpy==1.24.3
supabase==2.8.1
